이벤트 등록 함수 eventHandler 로 분리
